cleaned up unconstrained benchmarks : merged taichi with other benchs + saved bench data as npz + externalize plotting
