+++ analysis/unconstrained/plot_config.py
@@ -1,0 +1,20 @@
+LABELS = {'DDP': 'DDP',
+          'FDDP': 'FDDP (default LS)',
+          'FDDP_filter': 'FDDP (filter LS)',
+          'SQP': 'SQP',
+          }
+COLORS = {'DDP': 'r',
+          'FDDP': 'y',
+          'FDDP_filter': 'g',
+          'SQP': 'b',
+          }
+        
+LINESTYLES = {'DDP': 'dashdot',
+              'FDDP': 'dashed',
+              'FDDP_filter': 'dotted',
+              'SQP': 'solid',
+              }
+
+LABELSIZE = 22
+FONTSIZE  = 26
+FIGSIZE   = (19.2,10.8)
