+++ analysis/unconstrained/problems.py
@@ -1,0 +1,262 @@
+import example_robot_data
+import crocoddyl
+import pinocchio as pin
+import numpy as np
+from mim_robots.robot_loader import load_pinocchio_wrapper
+from crocoddyl.utils.pendulum import CostModelDoublePendulum, ActuationModelDoublePendulum
+from mim_robots.robot_loader import load_pinocchio_wrapper
+
+
+def create_double_pendulum_problem(x0):
+    '''
+    Create shooting problem for the double pendulum model
+    '''
+    print("Created double pendulum problem ...")
+    # Loading the double pendulum model
+    pendulum = example_robot_data.load('double_pendulum')
+    model = pendulum.model
+    state = crocoddyl.StateMultibody(model)
+    actuation = ActuationModelDoublePendulum(state, actLink=1)
+    nu = actuation.nu
+    runningCostModel = crocoddyl.CostModelSum(state, nu)
+    terminalCostModel = crocoddyl.CostModelSum(state, nu)
+    xResidual = crocoddyl.ResidualModelState(state, state.zero(), nu)
+    xActivation = crocoddyl.ActivationModelQuad(state.ndx)
+    uResidual = crocoddyl.ResidualModelControl(state, nu)
+    xRegCost = crocoddyl.CostModelResidual(state, xActivation, xResidual)
+    uRegCost = crocoddyl.CostModelResidual(state, uResidual)
+    xPendCost = CostModelDoublePendulum(state, crocoddyl.ActivationModelWeightedQuad(np.array([1.] * 4 + [0.1] * 2)), nu)
+    dt = 1e-2
+    runningCostModel.addCost("uReg", uRegCost, 1e-4 / dt)
+    runningCostModel.addCost("xGoal", xPendCost, 1e-5 / dt)
+    terminalCostModel.addCost("xGoal", xPendCost, 100.)
+    runningModel = crocoddyl.IntegratedActionModelEuler(
+        crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, runningCostModel), dt)
+    terminalModel = crocoddyl.IntegratedActionModelEuler(
+        crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, terminalCostModel), dt)
+    T = 100
+    pb = crocoddyl.ShootingProblem(x0, [runningModel] * T, terminalModel)
+    return pb
+
+def create_kuka_problem(x0):
+    '''
+    Create shooting problem for kuka reaching task
+    '''
+    print("Create kuka problem ...")
+    robot       = load_pinocchio_wrapper('iiwa')
+    model = robot.model
+    nq = model.nq; nv = model.nv
+    # State and actuation model
+    state = crocoddyl.StateMultibody(model)
+    actuation = crocoddyl.ActuationModelFull(state)
+    # Running and terminal cost models
+    runningCostModel = crocoddyl.CostModelSum(state)
+    terminalCostModel = crocoddyl.CostModelSum(state)
+    # Create cost terms 
+    # Control regularization cost
+    uResidual = crocoddyl.ResidualModelControlGrav(state)
+    uRegCost = crocoddyl.CostModelResidual(state, uResidual)
+    # State regularization cost
+    xResidual = crocoddyl.ResidualModelState(state, x0)
+    xRegCost = crocoddyl.CostModelResidual(state, xResidual)
+    # endeff frame translation cost
+    endeff_frame_id = model.getFrameId("contact")
+    # endeff_translation = robot.data.oMf[endeff_frame_id].translation.copy()
+    endeff_translation = np.array([0.7, 0, 1.1]) # move endeff +30 cm along x in WORLD frame
+    frameTranslationResidual = crocoddyl.ResidualModelFrameTranslation(state, endeff_frame_id, endeff_translation)
+    frameTranslationCost = crocoddyl.CostModelResidual(state, frameTranslationResidual)
+    # Add costs
+    runningCostModel.addCost("stateReg", xRegCost, 1e-1)
+    runningCostModel.addCost("ctrlRegGrav", uRegCost, 1e-4)
+    runningCostModel.addCost("translation", frameTranslationCost, 1)
+    terminalCostModel.addCost("stateReg", xRegCost, 1e-1)
+    terminalCostModel.addCost("translation", frameTranslationCost, 1)
+    # Create Differential Action Model (DAM), i.e. continuous dynamics and cost functions
+    running_DAM = crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, runningCostModel)
+    terminal_DAM = crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, terminalCostModel)
+    # Create Integrated Action Model (IAM), i.e. Euler integration of continuous dynamics and cost
+    dt = 1e-2
+    runningModel = crocoddyl.IntegratedActionModelEuler(running_DAM, dt)
+    terminalModel = crocoddyl.IntegratedActionModelEuler(terminal_DAM, 0.)
+    # Create the shooting problem
+    T = 50
+    pb = crocoddyl.ShootingProblem(x0, [runningModel] * T, terminalModel)
+    return pb
+
+def create_quadrotor_problem(x0):
+    '''
+    Create shooting problem for quadrotor task
+    '''
+    print("Create quadrotor problem ...")
+    hector = example_robot_data.load('hector')
+    robot_model = hector.model
+    target_pos = np.array([1., 0., 1.])
+    target_quat = pin.Quaternion(1., 0., 0., 0.)
+    state = crocoddyl.StateMultibody(robot_model)
+    d_cog, cf, cm, u_lim, l_lim = 0.1525, 6.6e-5, 1e-6, 5., 0.1
+    tau_f = np.array([[0., 0., 0., 0.], [0., 0., 0., 0.], [1., 1., 1., 1.], [0., d_cog, 0., -d_cog],
+                    [-d_cog, 0., d_cog, 0.], [-cm / cf, cm / cf, -cm / cf, cm / cf]])
+    actuation = crocoddyl.ActuationModelMultiCopterBase(state, tau_f)
+
+    nu = actuation.nu
+    runningCostModel = crocoddyl.CostModelSum(state, nu)
+    terminalCostModel = crocoddyl.CostModelSum(state, nu)
+    # Costs
+    xResidual = crocoddyl.ResidualModelState(state, state.zero(), nu)
+    xActivation = crocoddyl.ActivationModelWeightedQuad(np.array([0.1] * 3 + [1000.] * 3 + [1000.] * robot_model.nv))
+    uResidual = crocoddyl.ResidualModelControl(state, nu)
+    xRegCost = crocoddyl.CostModelResidual(state, xActivation, xResidual)
+    uRegCost = crocoddyl.CostModelResidual(state, uResidual)
+    goalTrackingResidual = crocoddyl.ResidualModelFramePlacement(state, robot_model.getFrameId("base_link"),
+                                                                pin.SE3(target_quat.matrix(), target_pos), nu)
+    goalTrackingCost = crocoddyl.CostModelResidual(state, goalTrackingResidual)
+    runningCostModel.addCost("xReg", xRegCost, 1e-6)
+    runningCostModel.addCost("uReg", uRegCost, 1e-6)
+    runningCostModel.addCost("trackPose", goalTrackingCost, 1e-2)
+    terminalCostModel.addCost("goalPose", goalTrackingCost, 3.)
+    dt = 3e-2
+    runningModel = crocoddyl.IntegratedActionModelEuler(
+        crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, runningCostModel), dt)
+    terminalModel = crocoddyl.IntegratedActionModelEuler(
+        crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, terminalCostModel), dt)
+
+    # Creating the shooting problem and the FDDP solver
+    T = 33
+    pb = crocoddyl.ShootingProblem(x0, [runningModel] * T, terminalModel)
+    return pb 
+
+def create_humanoid_taichi_problem(target):
+    '''
+    Create shooting problem for Talos taichi task
+    '''
+    print("Create humanoid problem ...")
+    # Load robot
+    robot = example_robot_data.load('talos')
+    rmodel = robot.model
+    # Create data structures
+    rdata = rmodel.createData()
+    state = crocoddyl.StateMultibody(rmodel)
+    actuation = crocoddyl.ActuationModelFloatingBase(state)
+    # Set integration time
+    DT = 5e-2
+    T = 40
+    # target = np.array([0.4, 0, 1.2])
+    # Initialize reference state, target and reference CoM
+    rightFoot = 'right_sole_link'
+    leftFoot = 'left_sole_link'
+    endEffector = 'gripper_left_joint'
+    endEffectorId = rmodel.getFrameId(endEffector)
+    rightFootId = rmodel.getFrameId(rightFoot)
+    leftFootId = rmodel.getFrameId(leftFoot)
+    q0 = rmodel.referenceConfigurations["half_sitting"]
+    x0reg = np.concatenate([q0, np.zeros(rmodel.nv)])
+    pin.forwardKinematics(rmodel, rdata, q0)
+    pin.updateFramePlacements(rmodel, rdata)
+    rfPos0 = rdata.oMf[rightFootId].translation
+    lfPos0 = rdata.oMf[leftFootId].translation
+    refGripper = rdata.oMf[rmodel.getFrameId("gripper_left_joint")].translation
+    comRef = (rfPos0 + lfPos0) / 2
+    comRef[2] = pin.centerOfMass(rmodel, rdata, q0)[2].item()
+    # Create two contact models used along the motion
+    contactModel1Foot = crocoddyl.ContactModelMultiple(state, actuation.nu)
+    contactModel2Feet = crocoddyl.ContactModelMultiple(state, actuation.nu)
+    supportContactModelLeft = crocoddyl.ContactModel6D(state, leftFootId, pin.SE3.Identity(), pin.LOCAL, actuation.nu,
+                                                    np.array([0, 40]))
+    supportContactModelRight = crocoddyl.ContactModel6D(state, rightFootId, pin.SE3.Identity(), pin.LOCAL, actuation.nu,
+                                                        np.array([0, 40]))
+    contactModel1Foot.addContact(rightFoot + "_contact", supportContactModelRight)
+    contactModel2Feet.addContact(leftFoot + "_contact", supportContactModelLeft)
+    contactModel2Feet.addContact(rightFoot + "_contact", supportContactModelRight)
+    # Cost for self-collision
+    maxfloat = sys.float_info.max
+    xlb = np.concatenate([ 
+        -maxfloat * np.ones(6),  # dimension of the SE(3) manifold
+        rmodel.lowerPositionLimit[7:],
+        -maxfloat * np.ones(state.nv)
+    ])
+    xub = np.concatenate([
+        maxfloat * np.ones(6),  # dimension of the SE(3) manifold
+        rmodel.upperPositionLimit[7:],
+        maxfloat * np.ones(state.nv)
+    ])
+    bounds = crocoddyl.ActivationBounds(xlb, xub, 1.)
+    xLimitResidual = crocoddyl.ResidualModelState(state, x0reg, actuation.nu)
+    xLimitActivation = crocoddyl.ActivationModelQuadraticBarrier(bounds)
+    limitCost = crocoddyl.CostModelResidual(state, xLimitActivation, xLimitResidual)
+    # Cost for state and control
+    xResidual = crocoddyl.ResidualModelState(state, x0reg, actuation.nu)
+    xActivation = crocoddyl.ActivationModelWeightedQuad(
+        np.array([0] * 3 + [10.] * 3 + [0.01] * (state.nv - 6) + [10] * state.nv)**2)
+    uResidual = crocoddyl.ResidualModelControl(state, actuation.nu)
+    xTActivation = crocoddyl.ActivationModelWeightedQuad(
+        np.array([0] * 3 + [10.] * 3 + [0.01] * (state.nv - 6) + [100] * state.nv)**2)
+    xRegCost = crocoddyl.CostModelResidual(state, xActivation, xResidual)
+    uRegCost = crocoddyl.CostModelResidual(state, uResidual)
+    xRegTermCost = crocoddyl.CostModelResidual(state, xTActivation, xResidual)
+    # Cost for target reaching: hand and foot
+    handTrackingResidual = crocoddyl.ResidualModelFramePlacement(state, endEffectorId, pin.SE3(np.eye(3), target),
+                                                                actuation.nu)
+    handTrackingActivation = crocoddyl.ActivationModelWeightedQuad(np.array([1] * 3 + [0.0001] * 3)**2)
+    handTrackingCost = crocoddyl.CostModelResidual(state, handTrackingActivation, handTrackingResidual)
+
+    footTrackingResidual = crocoddyl.ResidualModelFramePlacement(state, leftFootId,
+                                                                pin.SE3(np.eye(3), np.array([0., 0.4, 0.])),
+                                                                actuation.nu)
+    footTrackingActivation = crocoddyl.ActivationModelWeightedQuad(np.array([1, 1, 0.1] + [1.] * 3)**2)
+    footTrackingCost1 = crocoddyl.CostModelResidual(state, footTrackingActivation, footTrackingResidual)
+    footTrackingResidual = crocoddyl.ResidualModelFramePlacement(state, leftFootId,
+                                                                pin.SE3(np.eye(3), np.array([0.3, 0.15, 0.35])),
+                                                                actuation.nu)
+    footTrackingCost2 = crocoddyl.CostModelResidual(state, footTrackingActivation, footTrackingResidual)
+    # Cost for CoM reference
+    comResidual = crocoddyl.ResidualModelCoMPosition(state, comRef, actuation.nu)
+    comTrack = crocoddyl.CostModelResidual(state, comResidual)
+    # Create cost model per each action model. We divide the motion in 3 phases plus its terminal model
+    runningCostModel1 = crocoddyl.CostModelSum(state, actuation.nu)
+    runningCostModel2 = crocoddyl.CostModelSum(state, actuation.nu)
+    runningCostModel3 = crocoddyl.CostModelSum(state, actuation.nu)
+    terminalCostModel = crocoddyl.CostModelSum(state, actuation.nu)
+    # Then let's add the running and terminal cost functions
+    runningCostModel1.addCost("gripperPose", handTrackingCost, 1e2)
+    runningCostModel1.addCost("stateReg", xRegCost, 1e-3)
+    runningCostModel1.addCost("ctrlReg", uRegCost, 1e-4)
+    runningCostModel1.addCost("limitCost", limitCost, 1e3)
+
+    runningCostModel2.addCost("gripperPose", handTrackingCost, 1e2)
+    runningCostModel2.addCost("footPose", footTrackingCost1, 1e1)
+    runningCostModel2.addCost("stateReg", xRegCost, 1e-3)
+    runningCostModel2.addCost("ctrlReg", uRegCost, 1e-4)
+    runningCostModel2.addCost("limitCost", limitCost, 1e3)
+
+    runningCostModel3.addCost("gripperPose", handTrackingCost, 1e2)
+    runningCostModel3.addCost("footPose", footTrackingCost2, 1e1)
+    runningCostModel3.addCost("stateReg", xRegCost, 1e-3)
+    runningCostModel3.addCost("ctrlReg", uRegCost, 1e-4)
+    runningCostModel3.addCost("limitCost", limitCost, 1e3)
+
+    terminalCostModel.addCost("gripperPose", handTrackingCost, 1e2)
+    terminalCostModel.addCost("stateReg", xRegTermCost, 1e-3)
+    terminalCostModel.addCost("limitCost", limitCost, 1e3)
+
+    # Create the action model
+    dmodelRunning1 = crocoddyl.DifferentialActionModelContactFwdDynamics(state, actuation, contactModel2Feet,
+                                                                        runningCostModel1)
+    dmodelRunning2 = crocoddyl.DifferentialActionModelContactFwdDynamics(state, actuation, contactModel1Foot,
+                                                                        runningCostModel2)
+    dmodelRunning3 = crocoddyl.DifferentialActionModelContactFwdDynamics(state, actuation, contactModel1Foot,
+                                                                        runningCostModel3)
+    dmodelTerminal = crocoddyl.DifferentialActionModelContactFwdDynamics(state, actuation, contactModel1Foot,
+                                                                        terminalCostModel)
+
+    runningModel1 = crocoddyl.IntegratedActionModelEuler(dmodelRunning1, DT)
+    runningModel2 = crocoddyl.IntegratedActionModelEuler(dmodelRunning2, DT)
+    runningModel3 = crocoddyl.IntegratedActionModelEuler(dmodelRunning3, DT)
+    terminalModel = crocoddyl.IntegratedActionModelEuler(dmodelTerminal, 0)
+
+    # Problem definition
+    x0 = np.concatenate([q0, pin.utils.zero(state.nv)])
+    pb = crocoddyl.ShootingProblem(x0, [runningModel1] * T + [runningModel2] * T + [runningModel3] * T, terminalModel)
+    return pb
+
+
+
