+++ analysis/unconstrained/rollout_benchmark.py
@@ -12,182 +12,40 @@ Randomizing over initial states
 import numpy as np
 import crocoddyl
 import pinocchio as pin
-
-from robot_properties_kuka.config import IiwaConfig
+from problems import create_double_pendulum_problem, create_kuka_problem, create_humanoid_taichi_problem, create_quadrotor_problem
+from mim_robots.robot_loader import load_pinocchio_wrapper
 import example_robot_data
-# from bench_utils.cartpole_swingup import DifferentialActionModelCartpole
-from crocoddyl.utils.pendulum import CostModelDoublePendulum, ActuationModelDoublePendulum
 import mim_solvers
 import time
+from plot_config import LABELS, COLORS, LINESTYLES, LABELSIZE, FONTSIZE, FIGSIZE
 
-def create_double_pendulum_problem(x0):
-    '''
-    Create shooting problem for the double pendulum model
-    '''
-    print("Created double pendulum problem ...")
-    # Loading the double pendulum model
-    pendulum = example_robot_data.load('double_pendulum')
-    model = pendulum.model
-    state = crocoddyl.StateMultibody(model)
-    actuation = ActuationModelDoublePendulum(state, actLink=1)
-    nu = actuation.nu
-    runningCostModel = crocoddyl.CostModelSum(state, nu)
-    terminalCostModel = crocoddyl.CostModelSum(state, nu)
-    xResidual = crocoddyl.ResidualModelState(state, state.zero(), nu)
-    xActivation = crocoddyl.ActivationModelQuad(state.ndx)
-    uResidual = crocoddyl.ResidualModelControl(state, nu)
-    xRegCost = crocoddyl.CostModelResidual(state, xActivation, xResidual)
-    uRegCost = crocoddyl.CostModelResidual(state, uResidual)
-    xPendCost = CostModelDoublePendulum(state, crocoddyl.ActivationModelWeightedQuad(np.array([1.] * 4 + [0.1] * 2)), nu)
-    dt = 1e-2
-    runningCostModel.addCost("uReg", uRegCost, 1e-4 / dt)
-    runningCostModel.addCost("xGoal", xPendCost, 1e-5 / dt)
-    terminalCostModel.addCost("xGoal", xPendCost, 100.)
-    runningModel = crocoddyl.IntegratedActionModelEuler(
-        crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, runningCostModel), dt)
-    terminalModel = crocoddyl.IntegratedActionModelEuler(
-        crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, terminalCostModel), dt)
-    T = 100
-    pb = crocoddyl.ShootingProblem(x0, [runningModel] * T, terminalModel)
-    return pb
-
-def create_cartpole_problem(x0):
-    '''
-    Create shooting problem for Cartpole
-    '''
-    print("Create cartpole problem ...")
-    # Creating the DAM for the cartpole
-    cartpoleDAM = DifferentialActionModelCartpole()
-    # Using NumDiff for computing the derivatives. We specify the
-    # withGaussApprox=True to have approximation of the Hessian based on the
-    # Jacobian of the cost residuals.
-    cartpoleND = crocoddyl.DifferentialActionModelNumDiff(cartpoleDAM, True)
-    # Getting the IAM using the simpletic Euler rule
-    timeStep = 5e-2
-    cartpoleIAM = crocoddyl.IntegratedActionModelEuler(cartpoleND, timeStep)
-    # Creating the shooting problem
-    T = 50
-    terminalCartpole = DifferentialActionModelCartpole()
-    terminalCartpoleDAM = crocoddyl.DifferentialActionModelNumDiff(terminalCartpole, True)
-    terminalCartpoleIAM = crocoddyl.IntegratedActionModelEuler(terminalCartpoleDAM, 0.)
-    terminalCartpole.costWeights[0] = 200
-    terminalCartpole.costWeights[1] = 200
-    terminalCartpole.costWeights[2] = 1.
-    terminalCartpole.costWeights[3] = 0.1
-    terminalCartpole.costWeights[4] = 0.01
-    terminalCartpole.costWeights[5] = 0.0001
-    pb = crocoddyl.ShootingProblem(x0, [cartpoleIAM] * T, terminalCartpoleIAM)
-    return pb 
-
-def create_kuka_problem(x0):
-    '''
-    Create shooting problem for kuka reaching task
-    '''
-    print("Create kuka problem ...")
-    iiwa_config = IiwaConfig()
-    robot       = iiwa_config.buildRobotWrapper()
-    model = robot.model
-    nq = model.nq; nv = model.nv
-    # State and actuation model
-    state = crocoddyl.StateMultibody(model)
-    actuation = crocoddyl.ActuationModelFull(state)
-    # Running and terminal cost models
-    runningCostModel = crocoddyl.CostModelSum(state)
-    terminalCostModel = crocoddyl.CostModelSum(state)
-    # Create cost terms 
-    # Control regularization cost
-    uResidual = crocoddyl.ResidualModelControlGrav(state)
-    uRegCost = crocoddyl.CostModelResidual(state, uResidual)
-    # State regularization cost
-    xResidual = crocoddyl.ResidualModelState(state, x0)
-    xRegCost = crocoddyl.CostModelResidual(state, xResidual)
-    # endeff frame translation cost
-    endeff_frame_id = model.getFrameId("contact")
-    # endeff_translation = robot.data.oMf[endeff_frame_id].translation.copy()
-    endeff_translation = np.array([0.7, 0, 1.1]) # move endeff +30 cm along x in WORLD frame
-    frameTranslationResidual = crocoddyl.ResidualModelFrameTranslation(state, endeff_frame_id, endeff_translation)
-    frameTranslationCost = crocoddyl.CostModelResidual(state, frameTranslationResidual)
-    # Add costs
-    runningCostModel.addCost("stateReg", xRegCost, 1e-1)
-    runningCostModel.addCost("ctrlRegGrav", uRegCost, 1e-4)
-    runningCostModel.addCost("translation", frameTranslationCost, 1)
-    terminalCostModel.addCost("stateReg", xRegCost, 1e-1)
-    terminalCostModel.addCost("translation", frameTranslationCost, 1)
-    # Create Differential Action Model (DAM), i.e. continuous dynamics and cost functions
-    running_DAM = crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, runningCostModel)
-    terminal_DAM = crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, terminalCostModel)
-    # Create Integrated Action Model (IAM), i.e. Euler integration of continuous dynamics and cost
-    dt = 1e-2
-    runningModel = crocoddyl.IntegratedActionModelEuler(running_DAM, dt)
-    terminalModel = crocoddyl.IntegratedActionModelEuler(terminal_DAM, 0.)
-    # Create the shooting problem
-    T = 50
-    pb = crocoddyl.ShootingProblem(x0, [runningModel] * T, terminalModel)
-    return pb
-
-def create_quadrotor_problem(x0):
-    '''
-    Create shooting problem for quadrotor task
-    '''
-    print("Create quadrotor problem ...")
-    hector = example_robot_data.load('hector')
-    robot_model = hector.model
-    target_pos = np.array([1., 0., 1.])
-    target_quat = pin.Quaternion(1., 0., 0., 0.)
-    state = crocoddyl.StateMultibody(robot_model)
-    d_cog, cf, cm, u_lim, l_lim = 0.1525, 6.6e-5, 1e-6, 5., 0.1
-    tau_f = np.array([[0., 0., 0., 0.], [0., 0., 0., 0.], [1., 1., 1., 1.], [0., d_cog, 0., -d_cog],
-                    [-d_cog, 0., d_cog, 0.], [-cm / cf, cm / cf, -cm / cf, cm / cf]])
-    actuation = crocoddyl.ActuationModelMultiCopterBase(state, tau_f)
-
-    nu = actuation.nu
-    runningCostModel = crocoddyl.CostModelSum(state, nu)
-    terminalCostModel = crocoddyl.CostModelSum(state, nu)
-    # Costs
-    xResidual = crocoddyl.ResidualModelState(state, state.zero(), nu)
-    xActivation = crocoddyl.ActivationModelWeightedQuad(np.array([0.1] * 3 + [1000.] * 3 + [1000.] * robot_model.nv))
-    uResidual = crocoddyl.ResidualModelControl(state, nu)
-    xRegCost = crocoddyl.CostModelResidual(state, xActivation, xResidual)
-    uRegCost = crocoddyl.CostModelResidual(state, uResidual)
-    goalTrackingResidual = crocoddyl.ResidualModelFramePlacement(state, robot_model.getFrameId("base_link"),
-                                                                pin.SE3(target_quat.matrix(), target_pos), nu)
-    goalTrackingCost = crocoddyl.CostModelResidual(state, goalTrackingResidual)
-    runningCostModel.addCost("xReg", xRegCost, 1e-6)
-    runningCostModel.addCost("uReg", uRegCost, 1e-6)
-    runningCostModel.addCost("trackPose", goalTrackingCost, 1e-2)
-    terminalCostModel.addCost("goalPose", goalTrackingCost, 3.)
-    dt = 3e-2
-    runningModel = crocoddyl.IntegratedActionModelEuler(
-        crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, runningCostModel), dt)
-    terminalModel = crocoddyl.IntegratedActionModelEuler(
-        crocoddyl.DifferentialActionModelFreeFwdDynamics(state, actuation, terminalCostModel), dt)
-
-    # Creating the shooting problem and the FDDP solver
-    T = 33
-    pb = crocoddyl.ShootingProblem(x0, [runningModel] * T, terminalModel)
-    return pb 
-
-
-
+# Benchmark name 
+BENCH_NAME = 'Taichi'
+# BENCH_NAME must be :
+#  'Pendulum'  # maxiter = 500
+#  'Kuka'      # maxiter = 100
+#  'Quadrotor' # maxiter = 200
+#  'Taichi'    # maxiter = 300
+if(BENCH_NAME == 'Pendulum'):
+    MAXITER = 500 
+elif(BENCH_NAME == 'Kuka'):
+    MAXITER = 100 
+elif(BENCH_NAME == 'Quadrotor'):
+    MAXITER = 200 
+elif(BENCH_NAME == 'Taichi'):
+    MAXITER = 300 
+else:
+    print("Error: wrong bench name !")
 
 # Solver params
-MAXITER     = 200 
 TOL         = 1e-4 
 CALLBACKS   = False
 FILTER_SIZE = MAXITER
-SAVE        = True # Save figure 
-# REG         = 1e-6
+SAVE        = True  
 
 # Benchmark params
 SEED = 1 ; np.random.seed(SEED)
-N_samples = 100
-names = [
-    #    'Pendulum'] # maxiter = 500
-        #  'Kuka'] # maxiter = 100
-        # #  'Cartpole']  #--> need to explain why it doesn't converge otherwise leave it out 
-         'Quadrotor'] # maxiter = 200
-
-N_pb = len(names)
+N_SAMPLES = 100
 
 # Solvers
 solversDDP         = []
@@ -201,80 +59,76 @@ cartpole_x0  = np.array([0., 3.14, 0., 0.])
 kuka_x0      = np.array([0.1, 0.7, 0., 0.7, -0.5, 1.5, 0.] + [0.]*7)
 quadrotor    = example_robot_data.load('hector') 
 quadrotor_x0 = np.array(list(quadrotor.q0) + [0.]*quadrotor.model.nv) 
+humanoid_x0  = np.array([0.4, 0, 1.2])
 
 # Create 1 solver of each type for each problem
 print('------')
-for k,name in enumerate(names):
-    if(name == "Pendulum"):  
-        pb = create_double_pendulum_problem(pendulum_x0)
-    if(name == "Cartpole"):  
-        pb = create_cartpole_problem(cartpole_x0) 
-    if(name == "Kuka"):      
-        pb = create_kuka_problem(kuka_x0) 
-    if(name == "Quadrotor"): 
-        pb = create_quadrotor_problem(quadrotor_x0) 
+if(BENCH_NAME == "Pendulum"):  
+    pb = create_double_pendulum_problem(pendulum_x0)
+if(BENCH_NAME == "Kuka"):      
+    pb = create_kuka_problem(kuka_x0) 
+if(BENCH_NAME == "Quadrotor"): 
+    pb = create_quadrotor_problem(quadrotor_x0) 
+if(BENCH_NAME == "Taichi"): 
+    pb = create_humanoid_taichi_problem(humanoid_x0) 
 
-    # Create solver DDP (SS)
-    solverddp = mim_solvers.SolverDDP(pb)
-    solverddp.xs = [solverddp.problem.x0] * (solverddp.problem.T + 1)  
-    solverddp.us = solverddp.problem.quasiStatic([solverddp.problem.x0] * solverddp.problem.T)
-    solverddp.termination_tolerance = TOL
-    if(CALLBACKS): solverddp.setCallbacks([crocoddyl.CallbackVerbose()])
-    solversDDP.append(solverddp)
-    
-    # Create solver FDDP (MS)
-    solverfddp = mim_solvers.SolverFDDP(pb)
-    solverfddp.xs = [solverfddp.problem.x0] * (solverfddp.problem.T + 1)  
-    solverfddp.us = solverfddp.problem.quasiStatic([solverfddp.problem.x0] * solverfddp.problem.T)
-    solverfddp.termination_tolerance = TOL
-    solverfddp.use_filter_line_search = False
-    if(CALLBACKS): solverfddp.setCallbacks([crocoddyl.CallbackVerbose()])
-    solversFDDP.append(solverfddp)
+# Create solver DDP (SS)
+solverddp = mim_solvers.SolverDDP(pb)
+solverddp.xs = [solverddp.problem.x0] * (solverddp.problem.T + 1)  
+solverddp.us = solverddp.problem.quasiStatic([solverddp.problem.x0] * solverddp.problem.T)
+solverddp.termination_tolerance = TOL
+if(CALLBACKS): solverddp.setCallbacks([crocoddyl.CallbackVerbose()])
+solversDDP.append(solverddp)
 
-    # Create solver FDDP_filter (MS)
-    solverfddp_filter = mim_solvers.SolverFDDP(pb)
-    solverfddp_filter.xs = [solverfddp_filter.problem.x0] * (solverfddp_filter.problem.T + 1)  
-    solverfddp_filter.us = solverfddp_filter.problem.quasiStatic([solverfddp_filter.problem.x0] * solverfddp_filter.problem.T)
-    solverfddp_filter.termination_tolerance  = TOL
-    solverfddp_filter.use_filter_line_search = True
-    solverfddp_filter.filter_size            = MAXITER
-    if(CALLBACKS): solverfddp_filter.setCallbacks([crocoddyl.CallbackVerbose()])
-    solversFDDP_filter.append(solverfddp_filter)
+# Create solver FDDP (MS)
+solverfddp = mim_solvers.SolverFDDP(pb)
+solverfddp.xs = [solverfddp.problem.x0] * (solverfddp.problem.T + 1)  
+solverfddp.us = solverfddp.problem.quasiStatic([solverfddp.problem.x0] * solverfddp.problem.T)
+solverfddp.termination_tolerance = TOL
+solverfddp.use_filter_line_search = False
+if(CALLBACKS): solverfddp.setCallbacks([crocoddyl.CallbackVerbose()])
+solversFDDP.append(solverfddp)
 
-    # Create solver SQP (MS)
-    solverSQP = mim_solvers.SolverSQP(pb)
-    solverSQP.xs = [solverSQP.problem.x0] * (solverSQP.problem.T + 1)  
-    solverSQP.us = solverSQP.problem.quasiStatic([solverSQP.problem.x0] * solverSQP.problem.T)
-    solverSQP.termination_tolerance  = TOL
-    solverSQP.use_filter_line_search = True
-    solverSQP.filter_size            = MAXITER
-    solverSQP.with_callbacks         = CALLBACKS
-    # solverSQP.reg_incFactor = 1.01
-    # solverSQP.reg_decFactor = 1.01
-    # solverSQP.reg_min = REG
-    # solverSQP.reg_max = REG
-    # solverSQP.regMin  = REG
-    # solverSQP.regMax  = REG
-    solverCSSQP.append(solverSQP)
+# Create solver FDDP_filter (MS)
+solverfddp_filter = mim_solvers.SolverFDDP(pb)
+solverfddp_filter.xs = [solverfddp_filter.problem.x0] * (solverfddp_filter.problem.T + 1)  
+solverfddp_filter.us = solverfddp_filter.problem.quasiStatic([solverfddp_filter.problem.x0] * solverfddp_filter.problem.T)
+solverfddp_filter.termination_tolerance  = TOL
+solverfddp_filter.use_filter_line_search = True
+solverfddp_filter.filter_size            = MAXITER
+if(CALLBACKS): solverfddp_filter.setCallbacks([crocoddyl.CallbackVerbose()])
+solversFDDP_filter.append(solverfddp_filter)
 
+# Create solver SQP (MS)
+solverSQP = mim_solvers.SolverSQP(pb)
+solverSQP.xs = [solverSQP.problem.x0] * (solverSQP.problem.T + 1)  
+solverSQP.us = solverSQP.problem.quasiStatic([solverSQP.problem.x0] * solverSQP.problem.T)
+solverSQP.termination_tolerance  = TOL
+solverSQP.use_filter_line_search = True
+solverSQP.filter_size            = MAXITER
+solverSQP.with_callbacks         = CALLBACKS
+solverCSSQP.append(solverSQP)
 
 
 # Initial state samples
-pendulum_x0_samples  = np.zeros((N_samples, 4))
-cartpole_x0_samples  = np.zeros((N_samples, 4))
-iiwa_config          = IiwaConfig()
-kuka                 = iiwa_config.buildRobotWrapper()
-kuka_x0_samples      = np.zeros((N_samples, kuka.model.nq + kuka.model.nv))
+pendulum_x0_samples  = np.zeros((N_SAMPLES, 4))
+cartpole_x0_samples  = np.zeros((N_SAMPLES, 4))
+kuka                 = load_pinocchio_wrapper("iiwa")
+kuka_x0_samples      = np.zeros((N_SAMPLES, kuka.model.nq + kuka.model.nv))
 quadrotor            = example_robot_data.load('hector') 
 humanoid             = example_robot_data.load('talos')
-quadrotor_x0_samples = np.zeros((N_samples, quadrotor.model.nq + quadrotor.model.nv))
-for i in range(N_samples):
+quadrotor_x0_samples = np.zeros((N_SAMPLES, quadrotor.model.nq + quadrotor.model.nv))
+humanoid_x0_samples  = np.zeros((N_SAMPLES, 3))
+for i in range(N_SAMPLES):
     pendulum_x0_samples[i,:]  = np.array([np.pi*(2*np.random.rand()-1), 0., 0., 0.])
     cartpole_x0_samples[i,:]  = np.array([0., np.pi/2, 0., 0.])
     kuka_x0_samples[i,:]      = np.concatenate([pin.randomConfiguration(kuka.model), np.zeros(kuka.model.nv)])
     quadrotor_x0_samples[i,:] = np.concatenate([pin.randomConfiguration(quadrotor.model), np.zeros(quadrotor.model.nv)])
+    err = np.zeros(3)
+    err[2] = 2*np.random.rand(1) - 1
+    humanoid_x0_samples[i,:]  = np.array([0.4, 0, 1.2]) + 0.5*err
 
-print("Created "+str(N_samples)+" random initial states per model !")
+print("Created "+str(N_SAMPLES)+" random initial states per model !")
 
 # Solve problems for sample initial states
 ddp_iter_samples   = []  
@@ -297,116 +151,108 @@ SQP_kkt_samples    = []
 SQP_solved_samples = []
 SQP_avg_time_per_iter_samples = []
 
-for i in range(N_samples):
-    ddp_iter_samples.append([])
-    ddp_kkt_samples.append([])
-    ddp_solved_samples.append([])
-    ddp_avg_time_per_iter_samples.append([])
-
-    fddp_iter_samples.append([])
-    fddp_kkt_samples.append([])
-    fddp_solved_samples.append([])
-    fddp_avg_time_per_iter_samples.append([])
-
-    fddp_filter_iter_samples.append([])
-    fddp_filter_kkt_samples.append([])
-    fddp_filter_solved_samples.append([])
-    fddp_filter_avg_time_per_iter_samples.append([])
-
-    SQP_iter_samples.append([])
-    SQP_kkt_samples.append([])
-    SQP_solved_samples.append([])
-    SQP_avg_time_per_iter_samples.append([])
-
+# Solve the problem for each sample
+for i in range(N_SAMPLES):
     print("---")
-    print("Sample "+str(i+1)+'/'+str(N_samples))
-    for k,name in enumerate(names):
-        # Initial state
-        if(name == "Pendulum"):  x0 = pendulum_x0_samples[i,:]
-        if(name == "Cartpole"):  x0 = cartpole_x0_samples[i,:]
-        if(name == "Kuka"):      x0 = kuka_x0_samples[i,:]
-        if(name == "Quadrotor"): x0 = quadrotor_x0_samples[i,:]
+    print("Sample "+str(i+1)+'/'+str(N_SAMPLES))
+    # Initial state
+    if(BENCH_NAME == "Pendulum"):  x0 = pendulum_x0_samples[i,:]
+    if(BENCH_NAME == "Kuka"):      x0 = kuka_x0_samples[i,:]
+    if(BENCH_NAME == "Quadrotor"): x0 = quadrotor_x0_samples[i,:]
+    if(BENCH_NAME == "Taichi"):    x0 = humanoid_x0_samples[i,:]
 
-        # DDP (SS)
-        print("   Problem : "+name+" DDP")
-        solverddp = solversDDP[k]
+    # DDP (SS)
+    print("   Problem : "+BENCH_NAME+" DDP")
+    if(BENCH_NAME == 'Taichi'):
+        models = list(solverddp.problem.runningModels) + [solverddp.problem.terminalModel]
+        for m in models: m.differential.costs.costs["gripperPose"].cost.residual.reference = pin.SE3(np.eye(3), x0.copy())
+    else:
         solverddp.problem.x0 = x0
-        solverddp.xs = [x0] * (solverddp.problem.T + 1) 
-        solverddp.us = solverddp.problem.quasiStatic([x0] * solverddp.problem.T)
-        tic = time.time()
-        solverddp.solve(solverddp.xs, solverddp.us, MAXITER, False)
-        ddp_solve_time = time.time() - tic
-        solved = (solverddp.iter < MAXITER) and (solverddp.KKT < TOL)
-        ddp_solved_samples[i].append( solved )
-        print("   iter = "+str(solverddp.iter)+"  |  KKT = "+str(solverddp.KKT))
-        if(not solved): 
-            print("      FAILED !!!!")
-            ddp_iter_samples[i].append(MAXITER)
-        else:
-            ddp_iter_samples[i].append(solverddp.iter)
-        ddp_avg_time_per_iter_samples[i].append(ddp_solve_time/solverddp.iter)
-        ddp_kkt_samples[i].append(solverddp.KKT)
+    solverddp.xs = [solverddp.problem.x0] * (solverddp.problem.T + 1) 
+    solverddp.us = solverddp.problem.quasiStatic([solverddp.problem.x0] * solverddp.problem.T)
+    tic = time.time()
+    solverddp.solve(solverddp.xs, solverddp.us, MAXITER, False)
+    ddp_solve_time = time.time() - tic
+    solved = (solverddp.iter < MAXITER) and (solverddp.KKT < TOL)
+    ddp_solved_samples.append( solved )
+    print("   iter = "+str(solverddp.iter)+"  |  KKT = "+str(solverddp.KKT))
+    if(not solved): 
+        print("      FAILED !!!!")
+        ddp_iter_samples.append(MAXITER)
+    else:
+        ddp_iter_samples.append(solverddp.iter)
+    ddp_avg_time_per_iter_samples.append(ddp_solve_time/solverddp.iter)
+    ddp_kkt_samples.append(solverddp.KKT)
 
-        # FDDP (MS)
-        print("   Problem : "+name+" FDDP")
-        solverfddp = solversFDDP[k]
+    # FDDP (MS)
+    print("   Problem : "+BENCH_NAME+" FDDP")
+    if(BENCH_NAME == 'Taichi'):
+        models = list(solverfddp.problem.runningModels) + [solverfddp.problem.terminalModel]
+        for m in models: m.differential.costs.costs["gripperPose"].cost.residual.reference = pin.SE3(np.eye(3), x0.copy())
+    else:
         solverfddp.problem.x0 = x0
-        solverfddp.xs = [x0] * (solverfddp.problem.T + 1) 
-        solverfddp.us = solverfddp.problem.quasiStatic([x0] * solverfddp.problem.T)
-        tic = time.time()
-        solverfddp.solve(solverfddp.xs, solverfddp.us, MAXITER, False)
-        fddp_solve_time = time.time() - tic
-        solved = (solverfddp.iter < MAXITER) and (solverfddp.KKT < TOL)
-        fddp_solved_samples[i].append( solved )
-        print("   iter = "+str(solverfddp.iter)+"  |  KKT = "+str(solverfddp.KKT))
-        if(not solved): 
-            print("      FAILED !!!!")
-            fddp_iter_samples[i].append(MAXITER)
-        else:
-            fddp_iter_samples[i].append(solverfddp.iter)
-        fddp_avg_time_per_iter_samples[i].append(fddp_solve_time/solverfddp.iter)
-        fddp_kkt_samples[i].append(solverfddp.KKT)
+    solverfddp.xs = [solverfddp.problem.x0] * (solverfddp.problem.T + 1) 
+    solverfddp.us = solverfddp.problem.quasiStatic([solverfddp.problem.x0] * solverfddp.problem.T)
+    tic = time.time()
+    solverfddp.solve(solverfddp.xs, solverfddp.us, MAXITER, False)
+    fddp_solve_time = time.time() - tic
+    solved = (solverfddp.iter < MAXITER) and (solverfddp.KKT < TOL)
+    fddp_solved_samples.append( solved )
+    print("   iter = "+str(solverfddp.iter)+"  |  KKT = "+str(solverfddp.KKT))
+    if(not solved): 
+        print("      FAILED !!!!")
+        fddp_iter_samples.append(MAXITER)
+    else:
+        fddp_iter_samples.append(solverfddp.iter)
+    fddp_avg_time_per_iter_samples.append(fddp_solve_time/solverfddp.iter)
+    fddp_kkt_samples.append(solverfddp.KKT)
 
-        # FDDP filter (MS)
-        print("   Problem : "+name+" FDDP_filter")
-        solverfddp_filter = solversFDDP_filter[k]
+    # FDDP filter (MS)
+    print("   Problem : "+BENCH_NAME+" FDDP_filter")
+    if(BENCH_NAME == 'Taichi'):
+        models = list(solverfddp_filter.problem.runningModels) + [solverfddp_filter.problem.terminalModel]
+        for m in models: m.differential.costs.costs["gripperPose"].cost.residual.reference = pin.SE3(np.eye(3), x0.copy())
+    else:
         solverfddp_filter.problem.x0 = x0
-        solverfddp_filter.xs = [x0] * (solverfddp_filter.problem.T + 1) 
-        solverfddp_filter.us = solverfddp_filter.problem.quasiStatic([x0] * solverfddp_filter.problem.T)
-        tic = time.time()
-        solverfddp_filter.solve(solverfddp_filter.xs, solverfddp_filter.us, MAXITER, False)
-        fddp_filter_solve_time = time.time() - tic
-        solved = (solverfddp_filter.iter < MAXITER) and (solverfddp_filter.KKT < TOL)
-        fddp_filter_solved_samples[i].append( solved )
-        print("   iter = "+str(solverfddp_filter.iter)+"  |  KKT = "+str(solverfddp_filter.KKT))
-        if(not solved): 
-            print("      FAILED !!!!")
-            fddp_filter_iter_samples[i].append(MAXITER)
-        else:
-            fddp_filter_iter_samples[i].append(solverfddp_filter.iter)
-        fddp_filter_avg_time_per_iter_samples[i].append(fddp_filter_solve_time/solverfddp_filter.iter)
-        fddp_filter_kkt_samples[i].append(solverfddp_filter.KKT)
+    solverfddp_filter.xs = [solverfddp_filter.problem.x0] * (solverfddp_filter.problem.T + 1) 
+    solverfddp_filter.us = solverfddp_filter.problem.quasiStatic([solverfddp_filter.problem.x0] * solverfddp_filter.problem.T)
+    tic = time.time()
+    solverfddp_filter.solve(solverfddp_filter.xs, solverfddp_filter.us, MAXITER, False)
+    fddp_filter_solve_time = time.time() - tic
+    solved = (solverfddp_filter.iter < MAXITER) and (solverfddp_filter.KKT < TOL)
+    fddp_filter_solved_samples.append( solved )
+    print("   iter = "+str(solverfddp_filter.iter)+"  |  KKT = "+str(solverfddp_filter.KKT))
+    if(not solved): 
+        print("      FAILED !!!!")
+        fddp_filter_iter_samples.append(MAXITER)
+    else:
+        fddp_filter_iter_samples.append(solverfddp_filter.iter)
+    fddp_filter_avg_time_per_iter_samples.append(fddp_filter_solve_time/solverfddp_filter.iter)
+    fddp_filter_kkt_samples.append(solverfddp_filter.KKT)
 
-        # SQP        
-        print("   Problem : "+name+" SQP")
-        solverSQP = solverCSSQP[k]
+    # SQP        
+    print("   Problem : "+BENCH_NAME+" SQP")
+    if(BENCH_NAME == 'Taichi'):
+        models = list(solverSQP.problem.runningModels) + [solverSQP.problem.terminalModel]
+        for m in models: m.differential.costs.costs["gripperPose"].cost.residual.reference = pin.SE3(np.eye(3), x0.copy())
+    else:
         solverSQP.problem.x0 = x0
-        solverSQP.xs = [x0] * (solverSQP.problem.T + 1) 
-        solverSQP.us = solverSQP.problem.quasiStatic([x0] * solverSQP.problem.T)
-        tic = time.time()
-        solverSQP.solve(solverSQP.xs, solverSQP.us, MAXITER, False)
-        SQP_solve_time = time.time() - tic
-            # Check convergence
-        solved = (solverSQP.iter < MAXITER) and (solverSQP.KKT < TOL)
-        SQP_solved_samples[i].append( solved )
-        print("   iter = "+str(solverSQP.iter)+"  |  KKT = "+str(solverSQP.KKT))
-        if(not solved): 
-            print("      FAILED !!!!")
-            SQP_iter_samples[i].append(MAXITER)
-        else:
-            SQP_iter_samples[i].append(solverSQP.iter)
-        SQP_avg_time_per_iter_samples[i].append(SQP_solve_time/solverSQP.iter)
-        SQP_kkt_samples[i].append(solverSQP.KKT)
+    solverSQP.xs = [solverSQP.problem.x0] * (solverSQP.problem.T + 1) 
+    solverSQP.us = solverSQP.problem.quasiStatic([solverSQP.problem.x0] * solverSQP.problem.T)
+    tic = time.time()
+    solverSQP.solve(solverSQP.xs, solverSQP.us, MAXITER, False)
+    SQP_solve_time = time.time() - tic
+        # Check convergence
+    solved = (solverSQP.iter < MAXITER) and (solverSQP.KKT < TOL)
+    SQP_solved_samples.append( solved )
+    print("   iter = "+str(solverSQP.iter)+"  |  KKT = "+str(solverSQP.KKT))
+    if(not solved): 
+        print("      FAILED !!!!")
+        SQP_iter_samples.append(MAXITER)
+    else:
+        SQP_iter_samples.append(solverSQP.iter)
+    SQP_avg_time_per_iter_samples.append(SQP_solve_time/solverSQP.iter)
+    SQP_kkt_samples.append(solverSQP.KKT)
 
 # Compute solving time statistics
 ddp_mean_solve_time         = np.mean(np.array(ddp_avg_time_per_iter_samples))
@@ -426,107 +272,38 @@ print(" FDDP_LS  = " , fddp_filter_mean_solve_time ,  ' \xB1 ' , fddp_filter_std
 print(" SQP      = " , SQP_mean_solve_time         ,  ' \xB1 ' , SQP_std_solve_time)
 
 # Average fddp iters
-ddp_iter_solved = np.zeros((MAXITER, N_pb))
-fddp_iter_solved = np.zeros((MAXITER, N_pb))
-fddp_filter_iter_solved = np.zeros((MAXITER, N_pb))
-SQP_iter_solved = np.zeros((MAXITER, N_pb))
+ddp_iter_solved         = np.zeros(MAXITER)
+fddp_iter_solved        = np.zeros(MAXITER)
+fddp_filter_iter_solved = np.zeros(MAXITER)
+SQP_iter_solved         = np.zeros(MAXITER)
 
-for k,exp in enumerate(names):
-    # Count number of problems solved for each sample initial state 
-    for i in range(N_samples):
-        # For sample i of problem k , compare nb iter to max iter
-        ddp_iter_ik  = np.array(ddp_iter_samples)[i,k]
-        fddp_iter_ik = np.array(fddp_iter_samples)[i,k]
-        fddp_filter_iter_ik = np.array(fddp_filter_iter_samples)[i,k]
-        SQP_iter_ik = np.array(SQP_iter_samples)[i,k]
-        for j in range(MAXITER):
-            if(ddp_iter_ik < j): ddp_iter_solved[j,k] += 1
-            if(fddp_iter_ik < j): fddp_iter_solved[j,k] += 1
-            if(fddp_filter_iter_ik < j): fddp_filter_iter_solved[j,k] += 1
-            if(SQP_iter_ik < j): SQP_iter_solved[j,k] += 1
+# Count number of problems solved for each sample initial state 
+for i in range(N_SAMPLES):
+    # For sample i of problem k , compare nb iter to max iter
+    ddp_iter_ik  = np.array(ddp_iter_samples)[i]
+    fddp_iter_ik = np.array(fddp_iter_samples)[i]
+    fddp_filter_iter_ik = np.array(fddp_filter_iter_samples)[i]
+    SQP_iter_ik = np.array(SQP_iter_samples)[i]
+    for j in range(MAXITER):
+        if(ddp_iter_ik < j): ddp_iter_solved[j] += 1
+        if(fddp_iter_ik < j): fddp_iter_solved[j] += 1
+        if(fddp_filter_iter_ik < j): fddp_filter_iter_solved[j] += 1
+        if(SQP_iter_ik < j): SQP_iter_solved[j] += 1
 
-
-# Generate plot of number of iterations for each problem
-import matplotlib.pyplot as plt
-import matplotlib
-matplotlib.rcParams["pdf.fonttype"] = 42
-matplotlib.rcParams["ps.fonttype"] = 42
-
-# x-axis : max number of iterations
-xdata     = range(0,MAXITER)
-xdata2     = range(0,N_samples)
-for k in range(N_pb):
-    # Plot number of problem solved vs max number of iterations
-    fig0, ax0 = plt.subplots(1, 1, figsize=(19.2,10.8))
-    ax0.plot(xdata, ddp_iter_solved[:,k]/N_samples, color='r', linewidth=4, linestyle='dashdot', label='DDP') 
-    ax0.plot(xdata, fddp_iter_solved[:,k]/N_samples, color='y', linewidth=4, linestyle='dashed', label='FDDP (default LS)') 
-    ax0.plot(xdata, fddp_filter_iter_solved[:,k]/N_samples, color='g', linewidth=4, linestyle='dotted', label='FDDP (filter LS)') 
-    ax0.plot(xdata, SQP_iter_solved[:,k]/N_samples, color='b', linewidth=4, linestyle='solid', label='SQP') #marker='o', markerfacecolor='b', linestyle='-', markersize=12, markeredgecolor='k', alpha=1., label='SQP')
-    # Set axis and stuff
-    ax0.set_ylabel('Percentage of problems solved', fontsize=26)
-    ax0.set_xlabel('Max. number of iterations', fontsize=26)
-    ax0.set_ylim(-0.02, 1.02)
-    ax0.tick_params(axis = 'y', labelsize=22)
-    ax0.tick_params(axis = 'x', labelsize=22)
-    ax0.grid(True) 
-    # Legend 
-    handles0, labels0 = ax0.get_legend_handles_labels()
-    fig0.legend(handles0, labels0, loc='lower right', bbox_to_anchor=(0.902, 0.1), prop={'size': 26}) 
-    # Save, show , clean
-    if(SAVE):
-        fig0.savefig('/home/skleff/Desktop/TRO-SQP/data/rollout_benchmarks/bench_'+names[k]+'_SEED='+str(SEED)+'_MAXITER='+str(MAXITER)+'_TOL='+str(TOL)+'_NEW.pdf', bbox_inches="tight")
-
-    # Plot average computation times 
-    # fig1, ax1 = plt.subplots(1, 1, figsize=(19.2,10.8))
-    # ax1.plot(xdata2, ddp_avg_time_per_iter_samples, color='r', linewidth=4, label='DDP') 
-    # ax1.plot(xdata2, np.mean(ddp_avg_time_per_iter_samples)*np.ones(N_samples), color='r', linewidth=8, alpha=0.3, linestyle='-.') 
-    # ax1.plot(xdata2, fddp_avg_time_per_iter_samples, color='y', linewidth=4, label='FDDP (default LS)') 
-    # ax1.plot(xdata2, np.mean(fddp_avg_time_per_iter_samples)*np.ones(N_samples), color='y', linewidth=8, alpha=0.3, linestyle='-.') 
-    # ax1.plot(xdata2, fddp_filter_avg_time_per_iter_samples, color='g', linewidth=4, label='FDDP (filter LS)') 
-    # ax1.plot(xdata2, np.mean(fddp_filter_avg_time_per_iter_samples)*np.ones(N_samples), color='g', linewidth=8, alpha=0.3, linestyle='-.') 
-    # ax1.plot(xdata2, SQP_avg_time_per_iter_samples, color='b', linewidth=4, label='SQP') #marker='o', markerfacecolor='b', linestyle='-', markersize=12, markeredgecolor='k', alpha=1., label='SQP')
-    # ax1.plot(xdata2, np.mean(SQP_avg_time_per_iter_samples)*np.ones(N_samples), color='b', linewidth=8, alpha=0.3, linestyle='-.') #marker='o', markerfacecolor='b', linestyle='-', markersize=12, markeredgecolor='k', alpha=1., label='SQP')
-    # # Set axis and stuff
-    # ax1.set_ylabel('Avg time per iteration', fontsize=26)
-    # ax1.set_xlabel('Random problem', fontsize=26)
-    # # ax1.set_ylim(-0.02, 1.02)
-    # ax1.tick_params(axis = 'y', labelsize=22)
-    # ax1.tick_params(axis = 'x', labelsize=22)
-    # ax1.grid(True) 
-    # # Legend 
-    # handles1, labels1 = ax1.get_legend_handles_labels()
-    # fig1.legend(handles1, labels1, loc='lower right', bbox_to_anchor=(0.902, 0.1), prop={'size': 26}) 
-    # # Save, show , clean
-    # if(SAVE):
-    #     fig1.savefig('/home/skleff/Desktop/TRO-SQP/data/rollout_benchmarks/bench_'+names[k]+'_SEED='+str(SEED)+'_MAXITER='+str(MAXITER)+'_TOL='+str(TOL)+'AVG_TIME.pdf', bbox_inches="tight")
-
-
-
-# # Plot CV
-# fig1, ax1 = plt.subplots(1, 1, figsize=(19.2,10.8)) 
-# # Create bar plot
-# X = np.arange(N_pb)
-# b2 = ax1.bar(X, fddp_iter_avg, yerr=fddp_iter_std, color = 'g', width = 0.22, capsize=10, label='FDDP')
-# b3 = ax1.bar(X + 0.13, SQP_iter_avg, yerr=SQP_iter_std, color = 'b', width = 0.22, capsize=10, label='SQP')
-# # b1 = ax1.bar(X - 0.13, fddp_iter_avg, yerr=fddp_iter_std, color = 'r', width = 0.25, capsize=10, label='FDDP')
-# # b2 = ax1.bar(X + 0.13, SQP_iter_avg, yerr=SQP_iter_std, color = 'g', width = 0.25, capsize=10, label='SQP')
-# # Set axis and stuff
-# ax1.set_ylabel('Number of iterations', fontsize=26)
-# ax1.set_ylim(-10, MAXITER)
-# # ax1.set_yticks(X)
-# ax1.tick_params(axis = 'y', labelsize=22)
-# # ax1.set_xlabel('Experiment', fontsize=26)
-# ax1.set_xticks(X)
-# ax1.set_xticklabels(names, rotation='horizontal', fontsize=18)
-# ax1.tick_params(axis = 'x', labelsize = 22)
-# # ax1.set_title('Performance of SQP and FDDP', fontdict={'size': 26})
-# ax1.grid(True) 
-# # Legend 
-# handles1, labels1 = ax1.get_legend_handles_labels()
-# fig1.legend(handles1, labels1, loc='upper right', prop={'size': 26}) #, ncols=2)
-# Save, show , clean
-# fig0.savefig('/home/skleff/data_paper_CSSQP/bench_'+names[0]+'_SEED='+str(SEED)+'_MAXITER='+str(MAXITER)+'_TOL='+str(TOL)+'.png')
-
-
-plt.show()
-plt.close('all')
+# Save the benchmark data 
+file_name = "/home/skleff/SQP_REBUTAL_BENCH/"+BENCH_NAME 
+np.savez_compressed(file_name, 
+         N_SAMPLES=N_SAMPLES,
+         MAXITER=MAXITER,
+         ddp_iter_solved=ddp_iter_solved, 
+         fddp_iter_solved=fddp_iter_solved,
+         fddp_filter_iter_solved=fddp_filter_iter_solved,
+         SQP_iter_solved=SQP_iter_solved, 
+         ddp_mean_solve_time=ddp_mean_solve_time,
+         ddp_std_solve_time=ddp_std_solve_time,
+         fddp_mean_solve_time=fddp_mean_solve_time, 
+         fddp_std_solve_time=fddp_std_solve_time,
+         fddp_filter_mean_solve_time=fddp_filter_mean_solve_time,
+         fddp_filter_std_solve_time=fddp_filter_std_solve_time, 
+         SQP_mean_solve_time=SQP_mean_solve_time,
+         SQP_std_solve_time=SQP_std_solve_time)
